+++ setup/doc.go
@@ -14,7 +14,15 @@ one participant is honest. In fact, all the participants would need to collude t
 maliciously.
 In addition, the software used to run the procedure must be audited carefully, to avoid bugs.
 
-For the bls12-381 curve, the largest distributed ceremony has been run by the Ethereum Foundation
+For the BN254 curve, AlgoPlonk includes the parameters from the battle tested perpetual
+“powers-of-tau” ceremony used by projects such as Semaphore, Hermez, Torndado Cash and snarkjs.
+The parameters included in AlgoPlonk support circuits up to 2^17 (128k) constraints, but the
+ceremony provided parameters that can support up to 2^27 (128M) constraints which could be added.
+
+Learn more about the ceremony here:
+https://github.com/privacy-scaling-explorations/perpetualpowersoftau
+
+For the BLS12-381 curve, the largest distributed ceremony has been run by the Ethereum Foundation
 to implement Proto-danksharding (EIP-4844), with over 140,000 participants and rigorous auditing.
 AlgoPlonk uses these parameters to secure BLS12-381 proof systems which can support circuits up to
 2^14 (16,384) constraints.
@@ -23,15 +31,15 @@ Learn more about the ceremony here:
 https://ceremony.ethereum.org/
 https://github.com/ethereum/kzg-ceremony
 
-For the bn254 curve, AlgoPlonk includes the parameters from the battle tested perpetual
-“powers-of-tau” ceremony used by projects such as Semaphore, Hermez, Torndado Cash and snarkjs.
-The parameters included in AlgoPlonk support circuits up to 2^17 (128k) constraints, but the
-ceremony provided parameters that can support up to 2^27 (128M) constraints which could be added.
+Another BLS12-381 ceremony has been run by Dusk Network, extending the Zcash ceremony that had 88
+participants, with 15 additional participants, so it is more secure than the original Zcash alone,
+which is already widely trusted and battle-tested.
+It generated parameters that can support circuits up to 2^21 (2M) constraints.
 
 Learn more about the ceremony here:
-https://github.com/privacy-scaling-explorations/perpetualpowersoftau
+https://github.com/dusk-network/trusted-setup
 
 To audit the trusted parameters used in this library and verify they match the original source,
-refer to the doc.go file in the bls12-381 and bn254 subfolders.
+refer to the doc.go file in the respective ceremony subfolder.
 */
 package setup